requirements.txt
openai==0.27.8
langchain==0.0.229
langchainplus-sdk==0.0.20
python-dotenv==1.0.0
pytube
# 15.0.0 bug pytube.exceptions.RegexMatchError: get_throttling_function_name: could not find match for multiple
# Need to install from clone fixed version pytube https://github.com/pytube/pytube/issues/1678
pydub
tiktoken